Use the JIT Global & Thread-local state in PMAP.
PiperOrigin-RevId: 388696560
