+++ tensorflow/compiler/xla/python/pmap_lib.cc
@@ -199,7 +199,10 @@ py::object PmapFunction::Call(py::args args, py::kwargs kwargs) {
   }
 
   // Get dynamic argument signatures.
-  const bool jax_enable_x64 = GetEnableX64();
+  GlobalJitState& global_state = jax::GetGlobalState();
+  ThreadLocalJitState& tls = jax::GetLocalState();
+  const bool jax_enable_x64 = tls.enable_x64.value_or(global_state.enable_x64);
+  arguments.signature.jax_enable_x64 = jax_enable_x64;
   for (py::handle arg : arguments.flat_dynamic_args) {
     auto signature_or_error = xla::PyArgSignatureOfValue(arg, jax_enable_x64);
     if (!signature_or_error.ok()) {
@@ -208,6 +211,8 @@ py::object PmapFunction::Call(py::args args, py::kwargs kwargs) {
     arguments.signature.dynamic_arg_signatures.push_back(
         std::move(signature_or_error).ValueOrDie());
   }
+  arguments.signature.global_extra_jit_context = global_state.extra_jit_context;
+  arguments.signature.thread_local_extra_jit_context = tls.extra_jit_context;
 
   // Retrieve/Maybe add the executable to the cache.
   PmapCacheEntry* cache_entry = GetCacheEntryIfPresent(arguments.signature);
